ADD: Mock Data를 활용하여 여러 개의 피드 구현/ css 수정 필요
