+++ src/pages/jiyeon/Main/MainJiyeon.scss
@@ -234,6 +234,7 @@
   .feeds {
     width: 480px;
     margin-right: 20px;
+    flex-wrap: wrap;
   }
 
   .feedTop {
